kvs: introducing `api` crate for api related definitions
