+++ core/src/kvs/fdb/mod.rs
@@ -5,7 +5,6 @@ mod cnf;
 use crate::err::Error;
 use crate::kvs::Check;
 use crate::kvs::Key;
-use crate::kvs::Transactable;
 use crate::kvs::Val;
 use crate::vs::{u64_to_versionstamp, Versionstamp};
 use foundationdb::options::DatabaseOption;
@@ -158,7 +157,7 @@ impl Transaction {
 	}
 }
 
-impl Transactable for Transaction {
+impl crate::kvs::api::Transaction for Transaction {
 	/// Check if closed
 	fn closed(&self) -> bool {
 		self.done
