+++ core/src/kvs/rocksdb/mod.rs
@@ -6,7 +6,6 @@ use crate::err::Error;
 use crate::key::error::KeyCategory;
 use crate::kvs::Check;
 use crate::kvs::Key;
-use crate::kvs::Transactable;
 use crate::kvs::Val;
 use crate::vs::{try_to_u64_be, u64_to_versionstamp, Versionstamp};
 use futures::lock::Mutex;
@@ -169,7 +168,7 @@ impl Transaction {
 	}
 }
 
-impl Transactable for Transaction {
+impl crate::kvs::api::Transaction for Transaction {
 	/// Check if closed
 	fn closed(&self) -> bool {
 		self.done
