+++ core/src/kvs/surrealkv/mod.rs
@@ -4,7 +4,6 @@ use crate::err::Error;
 use crate::key::error::KeyCategory;
 use crate::kvs::Check;
 use crate::kvs::Key;
-use crate::kvs::Transactable;
 use crate::kvs::Val;
 use crate::vs::{try_to_u64_be, u64_to_versionstamp, Versionstamp};
 
@@ -100,7 +99,7 @@ impl Transaction {
 	}
 }
 
-impl Transactable for Transaction {
+impl crate::kvs::api::Transaction for Transaction {
 	/// Checks if the transaction is closed.
 	fn closed(&self) -> bool {
 		self.done
